Fix extract method logic
