+++ RefactorMCP.Tests/Split/TestBase.cs
@@ -1,0 +1,22 @@
+using System;
+using System.IO;
+
+namespace RefactorMCP.Tests;
+
+public abstract class TestBase : IDisposable
+{
+    protected static readonly string SolutionPath = TestUtilities.GetSolutionPath();
+    protected static readonly string ExampleFilePath = Path.Combine(Path.GetDirectoryName(SolutionPath)!, "RefactorMCP.Tests", "ExampleCode.cs");
+    protected const string TestOutputPath = "./RefactorMCP.Tests/TestOutput";
+
+    protected TestBase()
+    {
+        Directory.CreateDirectory(TestOutputPath);
+    }
+
+    public void Dispose()
+    {
+        if (Directory.Exists(TestOutputPath))
+            Directory.Delete(TestOutputPath, true);
+    }
+}
